Simplify configure tests for aspell
Don't bother looking for the binary, as we don't need it.

Assume that the include file and library are available to the compiler (use
suitable arguments to configure if not).

Checking for get_aspell_dict_info_list is sufficient to ensure we have a
sufficiently new version.

Make back-end optional (default is to check for it and not complain if not
found).